requirements.txt
requests==2.23.0
protobuf==3.15.1
pytest==6.2.4
pytest-dotenv==0.5.2
coverage==5.5
pytest-html==3.1.1
pytest-asyncio==0.15.1
httpx==0.18.2
SQLAlchemy==1.4.27
bcrypt==3.2.0
python-multipart==0.0.5
pyjwt==2.3.0
pandas==1.3.5
alembic==1.7.5
psycopg2==2.9.3
gunicorn==20.1.0
uvicorn==0.17.6
fastapi==0.75.0
